Импорт каталога

+++ inventory/src/main/kotlin/com/example/inventory/model/Product.kt
@@ -4,10 +4,13 @@ import com.mongodb.DBRef
 import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.mapping.Document
 
+/**
+ * Сущность продукта.
+ */
 @Document("products")
 class Product(
     @Id val id: Long,
     var title: String,
-    var price: Double,
+    var price: Int,
     var categories: List<DBRef>
 )
